chore: app does not intercept htlcs
